requirements.txt
anthropic>=0.49.0
requests>=2.31.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
markdown>=3.5.2
python-dotenv>=1.0.0
